requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
sqlmodel==0.0.22
python-multipart==0.0.20
psycopg[binary]==3.2.10
python-dotenv==1.0.1
alembic==1.14.0
